Fix issue with rate change and remote control scripts and stand alone scripts

+++ src/lib/handler/deohandler.cpp
@@ -134,7 +134,7 @@ void HereSphereHandler::readData()
         QString path = jsonObject["path"].toString();
         qint64 duration = jsonObject["duration"].toDouble() * 1000;
         qint64 currentTime = jsonObject["currentTime"].toDouble() * 1000;
-        float playbackSpeed = jsonObject["playbackSpeed"].toDouble() * 1.0;
+        double playbackSpeed = jsonObject["playbackSpeed"].toDouble() * 1.0;
         bool playing = jsonObject["playerState"].toInt() == 0; // 0 == true? right? I know...
         bool stopped = false;
 //        LogHandler::Debug("Deo path: "+path);
